Implement Masso Link 2.15 uploads and watch subfolders
docs/protocol.md was written by analyst agents from the Masso Link
binaries. This implementation was written by this session, and the
implementer agents it spawned, from docs/protocol.md alone; none of them
opened the binaries, disassembly, captures, or analysis notes.

The codec follows the v2.15 layouts: a 32-bit serial in the identity
reply, whose version string is read from its fixed offset and never past
byte 41, with Identity.MaxTools for the tool-table ceiling; a zeroed
tool-query trailer; a status file name bounded to bytes 17-49; the
variable-length upload-start request, with a directory for subfolder
uploads, ValidateUploadDir, and JoinUploadPath; the 0xF7 start result;
and the 0x0C upload-abort notification. Serials are uint32 everywhere
they flow.

Client.Upload takes a directory and implements the v2.15 transfer: the
start request resent every second for up to five seconds, 0xF7 accepted
only after a resend, up to two chunks in flight with the smoothed
round-trip retransmit timeout clamped to 40-250 ms, the 15 s give-up with
no ACK of any kind, and three 0x0C notifications 20 ms apart after an
acknowledged transfer fails. Status, tool, and upload replies from any IP
but the controller's are dropped, and each upload keeps to the
controller it started with. Tools stops at MaxTools. The socket binds
IPv4 only, so a port another program holds is reported as taken. The
simulator counts aborts and chunk requests and gains knobs for lost start
ACKs, lost chunk requests, and lost chunk ACKs.

Subfolders are always watched, as README.md describes: the watcher scans
recursively, skipping the top-level sent folder and hidden or "~$"
folders, including ones with the Windows Hidden attribute. The engine
uploads each file into the matching folder on the USB drive, keys it by
its relative path, and archives it to sent/ under the same subfolder. A
retried file whose name or folder is still invalid is rejected again, and
a file from a new watch folder no longer merges into an in-flight one
from the old folder. The integration suite uploads MLTEST4.NC into an
MLTEST folder.

+++ internal/app/app_test.go
@@ -73,7 +73,7 @@ func (f *fakeEngine) Run(ctx context.Context) error {
 }
 
 func (f *fakeEngine) Events() <-chan engine.Event { return f.events }
-func (*fakeEngine) SetSerial(uint16)              {}
+func (*fakeEngine) SetSerial(uint32)              {}
 func (*fakeEngine) SetWatchDir(string) error      { return nil }
 func (*fakeEngine) SetUploadWhileMachining(bool)  {}
 func (*fakeEngine) RefreshTools()                 {}
